file clean up for readability
