+++ config/actions/generateIndexes.js
@@ -1,34 +1,3 @@
-const fs = require('fs');
-const { execSync } = require('child_process');
-const path = require('path');
-const globby = require('globby');
-const rimraf = require('rimraf');
-
-function setupEntrypoints (dictionary, config) {
-  const propertyKeys = Object.keys(dictionary.properties);
-  const destinationPath = path.resolve(__dirname, `../${config.buildPath}`, 'design-tokens-js');
-  propertyKeys.forEach(key => {
-    rimraf.sync(`${destinationPath}/${key}`);
-    const data = {
-      main: "dist/design-tokens-js.cjs.js",
-      module: "dist/design-tokens-js.esm.js",
-      preconstruct: {
-        source:  `../src/${key}`,
-      }
-    };
-    execSync(`mkdir ${destinationPath}/${key}`);
-    fs.writeFileSync(`${destinationPath}/${key}/package.json`, JSON.stringify(data, null, 2))
-  });
-}
-
-function removeEntrypoints () {
-  const propertyKeys = Object.keys(dictionary.properties);
-  const destinationPath = path.resolve(__dirname, `..${config.buildPath}/design-tokens-js`);
-  propertyKeys.forEach(key => {
-    rimraf.sync(`${destinationPath}/${key}/*`);
-  });
-}
-
 function createReferences (fileExt, importPaths) {
   switch (fileExt) {
     case 'js':
@@ -61,7 +30,6 @@ function generateIndexFiles (dictionary, config) {
   const sortedImportPaths = destinationFiles.reduce((acc, curr) => {
     const [match] = curr.match(/\.[0-9a-z]+$/i);
     const fileExt = match.substring(1);
-    console.log(fileExt);
     const importPath = curr.split('/').pop();
     if (acc[fileExt]) {
       acc[fileExt].push(importPath);
@@ -71,12 +39,8 @@ function generateIndexFiles (dictionary, config) {
     return acc;
   }, {});
 
-  console.log(sortedImportPaths);
-
   const references = prepareFiles(sortedImportPaths).filter( i => i.file);
-  console.log(references);
   references.forEach(ref => {
-    console.log('EWD',ref);
     try {
       fs.writeFileSync(path.resolve(__dirname, ref.targetPath), ref.file);
     } catch (e) {
@@ -89,15 +53,9 @@ function removeIndexFiles (dictionary, config) {
   console.log(dictionary, config);
 }
 
-const actions = [{
-  name: 'generate_index_files',
-  do: generateIndexFiles,
-  undo: removeIndexFiles,
-},
-{
-  name: 'setup_entrypoints',
-  do: setupEntrypoints,
-  undo: removeEntrypoints,
-}];
 
-module.exports = actions;
+
+module.exports = {
+  generateIndexFiles,
+  removeIndexFiles,
+}
